# proto/datatest.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tmp/proto/datatest.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/eat": {
      "post": {
        "operationId": "fisheat",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/datatestfishmessagetwo"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/datatestfishmessageone"
            }
          }
        ],
        "tags": [
          "Serviceone"
        ]
      }
    },
    "/sleep": {
      "get": {
        "operationId": "eggsleep",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/datatestfishmessagetwo"
            }
          }
        },
        "parameters": [
          {
            "name": "fishfieldthree",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "boolean",
              "format": "boolean"
            }
          },
          {
            "name": "fishfieldfour",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "egg"
        ]
      }
    }
  },
  "definitions": {
    "datatestfishmessageone": {
      "type": "object",
      "properties": {
        "fishfieldone": {
          "type": "string"
        },
        "fishfieldtwo": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "datatestfishmessagetwo": {
      "type": "object",
      "properties": {
        "fishfieldthree": {
          "type": "array",
          "items": {
            "type": "boolean",
            "format": "boolean"
          }
        },
        "fishfieldfour": {
          "type": "string"
        }
      }
    }
  }
}